requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
deprecation==2.1.0
firebase-admin==6.3.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
importlib-metadata==6.11.0
install==1.3.5
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.7
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.1
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
streamlit-local-storage==0.0.12
streamlit_modal==0.1.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
